requirements.txt
anthropic>=0.49.0
python-telegram-bot>=20.7
aiosqlite>=0.19.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
